requirements.txt
alabaster==0.7.12
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
packaging==19.1
Pillow==6.1.0
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.1
requests==2.22.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
urllib3==1.25.3
